uttertuple-new/authtuple/requirements.txt
alembic>=1.10.3
asyncpg>=0.27.0
authlib>=1.2.0
bcrypt>=4.0.1
boto3>=1.28.0
cryptography>=40.0.0
email-validator>=2.0.0
fastapi>=0.95.0
httpx>=0.24.0
passlib>=1.7.4
psycopg2-binary>=2.9.6
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyjwt>=2.6.0
pymysql>=1.0.3
python-dotenv>=1.0.0
python-jose>=3.3.0
python-multipart>=0.0.6
requests>=2.31.0
rich>=13.0.0 
sqlalchemy>=2.0.0
tenacity>=8.2.2
uvicorn>=0.21.1